[UUID] Change how the UUID is calculated and handled
The UUID struct is just an opaque identifier so it makes no sense for
it to care about endianness nor does it make sense to use the current
time as part of that unique identifier as compared to a 'normal' random
number. As per dbus specification, uuid section, it's perfectly fine to
use 128 bits of random data to generate a UUID, so that's what we do.
We also don't require unsafe code anymore.