Fix interface walking in mapper. Enable remaining unit tests.
Previously we were using GetInterfaces which return a flat hierarchy of all the interface implemented by a type. We then return the MemberInfo of each of those interface.

Problem is in cases like:

[Interface (...)]
interface IMain : IPrivate {}
interface IPrivate { void Method (); }

We would not return a MemberInfo for Method for IMain (the interface we are interested in) but for IPrivate which has no existence for D-Bus (the method call is sent the interface registered on IMain, not IPrivate). So we now walk the tree recursively for each of the interface to implement proper inheritance for GetPublicMembers which fixes the problem.