implementor: Correct F# Event Identification
The MethodInfo.IsSpecialName property doesn't appear to be set true for
event add/remove methods produced with the F# compiler, even with the
[<CLIEvent>] annotation applied to the event.

This appears partly due to F#'s generalisation of the signature for
event handler delegate types, but also because we attempt to treat all
methods equally when implementing them.

This patch provides a quick hack to solve the problem with minimal
change to the TypeImplementor class.